refactor rightpanel
