+++ src/pages/simulationpage/RightPanel.jsx
@@ -1,11 +1,9 @@
 import React, { Suspense } from 'react';
-const MonacoEditor = React.lazy(() => import('@monaco-editor/react'));
-const MonacoDiffEditor = React.lazy(() => import('@monaco-editor/react').then(m => ({ default: m.DiffEditor })));
+import RightPanelEditor from './components/RightPanelEditor';
+import { PlotterManager } from './components/PlotterManager';
+import { SerialTabBar, SerialOutputPane, SerialSendRow } from './components/SerialMonitor';
 import { Btn } from './Btn';
 import { getBoardColors } from './projectUtils';
-import PlotterCanvas from './PlotterCanvas';
-
-const PLOTTER_COLORS = ['#e74c3c', '#3498db', '#2ecc71', '#f1c40f', '#9b59b6', '#e67e22', '#1abc9c'];
 
 
 
@@ -15,7 +13,7 @@ const BlocklyEditor = React.lazy(() => import('../../components/BlocklyEditor.js
 const DISABLED_FILE_SUFFIX = '.disabled';
 
 
-function RightPanelInternal(props) {
+const RightPanelInternal = React.forwardRef((props, ref) => {
 
   const {
     isPanelOpen, panelWidth, isDragging, onMouseDownResize, setIsPanelOpen,
@@ -60,27 +58,15 @@ function RightPanelInternal(props) {
     serialBoardFilter && serialBoardFilter !== 'all' ? serialBoardFilter : 'all'
   );
   const [showSendTargetMenu, setShowSendTargetMenu] = React.useState(false);
-  const [localCode, setLocalCode] = React.useState(code);
-  const isInternalUpdate = React.useRef(false);
   const [compareWithId, setCompareWithId] = React.useState(null); // ID of the file to compare with active file
 
-  // Sync localCode with prop code when it changes from outside
-  React.useEffect(() => {
-    if (isInternalUpdate.current) {
-      isInternalUpdate.current = false;
-      return;
-    }
-    setLocalCode(code);
-  }, [code]);
+  const asideRef = React.useRef(null);
+  const explorerRef = React.useRef(null);
 
-  // Debounced update to central state
-  React.useEffect(() => {
-    if (localCode === code) return;
-    const timeout = setTimeout(() => {
-      setCode(localCode);
-    }, 400);
-    return () => clearTimeout(timeout);
-  }, [localCode, setCode, code]);
+  React.useImperativeHandle(ref, () => ({
+    get aside() { return asideRef.current; },
+    get explorer() { return explorerRef.current; }
+  }));
 
   const editorRef = React.useRef(null);
 
@@ -384,9 +370,10 @@ function RightPanelInternal(props) {
 
   return (
     <aside
+      ref={asideRef}
       className="relative bg-[var(--bg2)] border-l border-[var(--border)] flex flex-col shrink-0 overflow-hidden"
       style={{
-        width: isPanelOpen ? panelWidth : 21,
+        width: isDragging ? undefined : (isPanelOpen ? panelWidth : 21),
         transition: isDragging ? 'none' : 'width 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
         willChange: isDragging ? 'width' : 'auto',
         contain: isDragging ? 'size layout paint' : 'none'
@@ -660,7 +647,11 @@ function RightPanelInternal(props) {
                 <div style={{ display: 'flex', minHeight: 0, flex: 1 }}>
                   {showCodeExplorer && (
                     <>
-                      <div style={{ width: explorerWidth, borderRight: theme === 'light' ? '1px solid #cbd5e1' : '1px solid #1e2d47', display: 'flex', flexDirection: 'column', background: theme === 'light' ? '#f1f5f9' : '#090e1a', flexShrink: 0 }}>
+                      <div 
+                        ref={explorerRef}
+                        className="code-explorer-container"
+                        style={{ width: isExplorerDragging ? undefined : explorerWidth, borderRight: theme === 'light' ? '1px solid #cbd5e1' : '1px solid #1e2d47', display: 'flex', flexDirection: 'column', background: theme === 'light' ? '#f1f5f9' : '#090e1a', flexShrink: 0 }}
+                      >
                         <div style={{
                           padding: '10px 12px',
                           fontSize: 11,
@@ -1118,39 +1109,17 @@ function RightPanelInternal(props) {
                           </div>
                         );
                       })()}
-                      <Suspense fallback={<div style={{ padding: 16, color: 'var(--text3)' }}>Loading Editor...</div>}>
-                        {compareWithId ? (
-                          <div style={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
-                            <div style={{ padding: '4px 12px', background: 'var(--accent)', color: '#000', fontSize: 10, fontWeight: 800, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                              <span>COMPARING: {activeFile?.name} vs {(projectFiles || []).find(f => f.id === compareWithId)?.name}</span>
-                              <button onClick={() => setCompareWithId(null)} style={{ background: 'transparent', border: 'none', cursor: 'pointer', fontWeight: 800 }}>EXIT COMPARE</button>
-                            </div>
-                            <MonacoDiffEditor
-                              height="calc(100% - 20px)"
-                              original={(projectFiles || []).find(f => f.id === compareWithId)?.content || ''}
-                              modified={localCode}
-                              language={editorLanguage}
-                              theme={theme === 'light' ? 'openhw-light' : 'openhw-dark'}
-                              options={{ ...editorOptions, readOnly: true }}
-                            />
-                          </div>
-                        ) : (
-                          <MonacoEditor
-                            height="100%"
-                            language={editorLanguage}
-                            theme={theme === 'light' ? 'openhw-light' : 'openhw-dark'}
-                            value={localCode}
-                            onMount={handleEditorMount}
-                            onChange={v => {
-                              if (!activeCodeFileId || activeCodeFileId === 'project/diagram.json') return;
-                              if (editingDisabled) return;
-                              isInternalUpdate.current = true;
-                              setLocalCode(v || '');
-                            }}
-                            options={editorOptions}
-                          />
-                        )}
-                      </Suspense>
+                      <RightPanelEditor
+                        theme={theme}
+                        editorLanguage={editorLanguage}
+                        editorOptions={editorOptions}
+                        activeCodeFileId={activeCodeFileId}
+                        compareWithId={compareWithId}
+                        projectFiles={projectFiles}
+                        onSaveCodeFile={onSaveCodeFile}
+                        editingDisabled={editingDisabled}
+                        isDragging={isDragging}
+                      />
                     </div>
                   </div>
                 </div>
@@ -1586,94 +1555,23 @@ function RightPanelInternal(props) {
                     </div>
                   ) : (
                     /* Plotter Mode - Modernized with Per-Board Management */
-                    <div style={{ flex: 1, display: 'flex', flexDirection: 'column', overflow: 'hidden', position: 'relative' }}>
-                      <PlotterToolbar
-                        onAddChannel={() => setShowAddChannel(!showAddChannel)}
-                        isPaused={plotterPaused}
-                        onTogglePause={() => setPlotterPaused(!plotterPaused)}
-                        onClear={() => { if (plotDataRef.current) plotDataRef.current = []; }}
-                        timeDiv={plotterTimeDiv}
-                        setTimeDiv={setPlotterTimeDiv}
-                      />
-
-                      {showAddChannel && (
-                        <AddChannelPanel
-                          boardOptions={serialBoardOptions}
-                          boardLabels={serialBoardLabels}
-                          boardKinds={serialBoardKinds}
-                          boardColors={boardColors}
-                          selectedPins={selectedPlotPins}
-                          setSelectedPins={setSelectedPlotPins}
-                          onClose={() => setShowAddChannel(false)}
-                        />
-                      )}
-
-                      {/* Plotter Layout - Scrollable Container */}
-                      <div className="panel-scroll" style={{ flex: 1, display: 'flex', flexDirection: 'column', overflowY: 'auto', overflowX: 'hidden', position: 'relative' }}>
-                        <div style={{ display: 'flex', minHeight: '100%', width: '100%' }}>
-                          {/* Label Column */}
-                          {selectedPlotPins.length > 0 && (
-                            <div style={{
-                              width: 65,
-                              background: 'var(--bg2)',
-                              borderRight: '1px solid var(--border)',
-                              display: 'flex',
-                              flexDirection: 'column',
-                              flexShrink: 0,
-                              zIndex: 2
-                            }}>
-                              {selectedPlotPins.map((chan, i) => {
-                                const color = PLOTTER_COLORS[i % PLOTTER_COLORS.length];
-                                const boardLabel = serialBoardLabels[chan.boardId] || chan.boardId;
-                                return (
-                                  <div key={`${chan.boardId}:${chan.pinId}`} style={{
-                                    height: 80,
-                                    display: 'flex',
-                                    flexDirection: 'column',
-                                    alignItems: 'center',
-                                    justifyContent: 'center',
-                                    borderBottom: '1px solid var(--border)',
-                                    padding: '4px 2px',
-                                    gap: 3,
-                                    position: 'relative',
-                                    background: i % 2 === 1 ? 'rgba(255,255,255,0.01)' : 'transparent',
-                                    boxSizing: 'border-box'
-                                  }}>
-                                    <span style={{ fontSize: 9, color: 'var(--text4)', textTransform: 'lowercase', whiteSpace: 'nowrap', overflow: 'hidden', textOverflow: 'ellipsis', width: '100%', textAlign: 'center' }} title={boardLabel}>
-                                      {boardLabel}
-                                    </span>
-                                    <span style={{ fontSize: 11, fontWeight: 800, color: color, fontFamily: 'JetBrains Mono, monospace' }}>
-                                      {chan.pinId}
-                                    </span>
-                                    <button
-                                      onClick={() => setSelectedPlotPins(prev => prev.filter(p => p.boardId !== chan.boardId || p.pinId !== chan.pinId))}
-                                      style={{ background: 'transparent', border: 'none', color: 'var(--text4)', cursor: 'pointer', padding: 2, display: 'flex', alignItems: 'center' }}
-                                      className="hover:text-[var(--red)] transition-colors"
-                                      title="Remove channel"
-                                    >
-                                      <svg width="11" height="11" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><line x1="18" y1="6" x2="6" y2="18" /><line x1="6" y1="6" x2="18" y2="18" /></svg>
-                                    </button>
-                                  </div>
-                                );
-                              })}
-                            </div>
-                          )}
-
-                          {/* Canvas Area */}
-                          <PlotterCanvas
-                            plotDataRef={plotDataRef}
-                            selectedPlotPins={selectedPlotPins}
-                            plotterPaused={plotterPaused}
-                            plotterTimeDiv={plotterTimeDiv}
-                            theme={theme}
-                            isRunning={isRunning}
-                          />
-
-                        </div>
-                      </div>
-
-
-                    </div>
+                    <PlotterManager
+                      showAddChannel={showAddChannel}
+                      setShowAddChannel={setShowAddChannel}
+                      plotterPaused={plotterPaused}
+                      setPlotterPaused={setPlotterPaused}
+                      plotDataRef={plotDataRef}
+                      selectedPlotPins={selectedPlotPins}
+                      setSelectedPlotPins={setSelectedPlotPins}
+                      plotterTimeDiv={plotterTimeDiv}
+                      setPlotterTimeDiv={setPlotterTimeDiv}
+                      serialBoardOptions={serialBoardOptions}
+                      serialBoardLabels={serialBoardLabels}
+                      serialBoardKinds={serialBoardKinds}
+                      boardColors={boardColors}
+                      theme={theme}
+                      isRunning={isRunning}
+                    />
                   )}
                 </div>
               </div>
@@ -1683,41 +1581,42 @@ function RightPanelInternal(props) {
       )}
     </aside>
   );
-}
-const RightPanelBase = ({
-  isPanelOpen, panelWidth, isDragging, onMouseDownResize, setIsPanelOpen,
-  explorerWidth, isExplorerDragging, onMouseDownExplorerResize,
-  selected, setSelected, theme, projectName,
-  validationErrors, showValidation, setShowValidation,
-  healthScore, applyFix,
-  codeTab, setCodeTab, code, setCode,
-  blocklyXml, setBlocklyXml,
-  blocklyGeneratedCode, setBlocklyGeneratedCode,
-  useBlocklyCode, setUseBlocklyCode,
-  blocklyDisabled, setBlocklyDisabled,
-  projectFiles, openCodeTabs, activeCodeFileId, showCodeExplorer, onToggleCodeExplorer, onOpenCodeFile, onCloseCodeTab,
-  onSaveCodeFile, onDuplicateCodeFile, onRenameCodeFile, onDeleteCodeFile, onDownloadCodeFile,
-  onToggleCodeFileDisabled,
-  onCreateCodeFile, onCreateCodeTab, onUploadCodeFile,
-  libQuery, setLibQuery, handleSearchLibraries, isSearchingLib, libMessage, libInstalled, libResults, handleInstallLibrary, installingLib,
-  serialPaused, setSerialPaused, isRunning, serialHistory, setSerialHistory, serialOutputRef, serialInput, setSerialInput, sendSerialInput, clearSerialMonitor,
-  serialViewMode, setSerialViewMode, serialBoardFilter, setSerialBoardFilter, serialBoardOptions, serialBoardLabels, serialBoardKinds, serialBoardSourceModes, serialBaudRate, setSerialBaudRate, serialBaudOptions, serialLineEnding, setSerialLineEnding,
-  hardwareConnected,
-  plotterPaused, setPlotterPaused, plotDataRef, selectedPlotPins, setSelectedPlotPins, serialPlotLabelsRef,
-  showConnectionsPanel, wires, updateWireColor, deleteWire,
-  boardComponentMap, onToggleBoardFirmwareSource,
-  editingDisabled,
-  editingDisabledMessage,
-  boardLineEndings, setBoardLineEndings,
-  boardAutoscrolls, setBoardAutoscrolls,
-  boardBaudRates, setBoardBaudRates,
-  boardPausedStates, setBoardPausedStates,
-  boardInputs, setBoardInputs,
-  isSerialSplit, setIsSerialSplit,
-  serialSplitRatio, setSerialSplitRatio,
-  serialBoardFilter2, setSerialBoardFilter2,
-  plotterTimeDiv, setPlotterTimeDiv
-}) => {
+});
+const RightPanelBase = React.forwardRef((props, ref) => {
+  const {
+    isPanelOpen, panelWidth, isDragging, onMouseDownResize, setIsPanelOpen,
+    explorerWidth, isExplorerDragging, onMouseDownExplorerResize,
+    selected, setSelected, theme, projectName,
+    validationErrors, showValidation, setShowValidation,
+    healthScore, applyFix,
+    codeTab, setCodeTab, code, setCode,
+    blocklyXml, setBlocklyXml,
+    blocklyGeneratedCode, setBlocklyGeneratedCode,
+    useBlocklyCode, setUseBlocklyCode,
+    blocklyDisabled, setBlocklyDisabled,
+    projectFiles, openCodeTabs, activeCodeFileId, showCodeExplorer, onToggleCodeExplorer, onOpenCodeFile, onCloseCodeTab,
+    onSaveCodeFile, onDuplicateCodeFile, onRenameCodeFile, onDeleteCodeFile, onDownloadCodeFile,
+    onToggleCodeFileDisabled,
+    onCreateCodeFile, onCreateCodeTab, onUploadCodeFile,
+    libQuery, setLibQuery, handleSearchLibraries, isSearchingLib, libMessage, libInstalled, libResults, handleInstallLibrary, installingLib,
+    serialPaused, setSerialPaused, isRunning, serialHistory, setSerialHistory, serialOutputRef, serialInput, setSerialInput, sendSerialInput, clearSerialMonitor,
+    serialViewMode, setSerialViewMode, serialBoardFilter, setSerialBoardFilter, serialBoardOptions, serialBoardLabels, serialBoardLabels: serialBoardLabelsProp, serialBoardKinds, serialBoardSourceModes, serialBaudRate, setSerialBaudRate, serialBaudOptions, serialLineEnding, setSerialLineEnding,
+    hardwareConnected,
+    plotterPaused, setPlotterPaused, plotDataRef, selectedPlotPins, setSelectedPlotPins, serialPlotLabelsRef,
+    showConnectionsPanel, wires, updateWireColor, deleteWire,
+    boardComponentMap, onToggleBoardFirmwareSource,
+    editingDisabled,
+    editingDisabledMessage,
+    boardLineEndings, setBoardLineEndings,
+    boardAutoscrolls, setBoardAutoscrolls,
+    boardBaudRates, setBoardBaudRates,
+    boardPausedStates, setBoardPausedStates,
+    boardInputs, setBoardInputs,
+    isSerialSplit, setIsSerialSplit,
+    serialSplitRatio, setSerialSplitRatio,
+    serialBoardFilter2, setSerialBoardFilter2,
+    plotterTimeDiv, setPlotterTimeDiv
+  } = props;
 
   const [activePanel, setActivePanel] = React.useState('code');
 
@@ -1728,6 +1627,7 @@ const RightPanelBase = ({
 
   return (
     <RightPanelInternal
+      ref={ref}
       {...{
         isPanelOpen, panelWidth, isDragging, onMouseDownResize, setIsPanelOpen,
         explorerWidth, isExplorerDragging, onMouseDownExplorerResize,
@@ -1764,775 +1664,12 @@ const RightPanelBase = ({
       }}
     />
   );
-};
+});
 
-const PlotterToolbar = ({ onAddChannel, isPaused, onTogglePause, onClear, timeDiv, setTimeDiv }) => {
-  return (
-    <div style={{ padding: '6px 12px', borderBottom: '1px solid var(--border)', display: 'flex', gap: 10, alignItems: 'center', background: 'var(--bg2)', flexShrink: 0 }}>
-      <button
-        onClick={onAddChannel}
-        style={{
-          background: 'var(--accent)',
-          color: '#000',
-          border: 'none',
-          borderRadius: 6,
-          padding: '4px 12px',
-          fontSize: 11,
-          fontWeight: 700,
-          cursor: 'pointer',
-          display: 'flex',
-          alignItems: 'center',
-          gap: 6
-        }}
-        className="hover:brightness-110 active:scale-95 transition-all"
-      >
-        <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" /></svg>
-        Add Channel
-      </button>
 
-      <div style={{ display: 'flex', alignItems: 'center', gap: 6, marginLeft: 'auto' }}>
-        <span style={{ fontSize: 10, color: 'var(--text3)', fontWeight: 600 }}>Window:</span>
-        <select
-          value={timeDiv}
-          onChange={(e) => setTimeDiv(Number(e.target.value))}
-          style={{
-            background: 'var(--bg3)',
-            border: '1px solid var(--border)',
-            color: 'var(--text2)',
-            borderRadius: 4,
-            fontSize: 10,
-            padding: '2px 4px',
-            outline: 'none',
-            cursor: 'pointer'
-          }}
-        >
-          <option value={1}>1ms</option>
-          <option value={10}>10ms</option>
-          <option value={100}>100ms</option>
-          <option value={500}>500ms</option>
-          <option value={1000}>1000ms</option>
-          <option value={2000}>2000ms</option>
-        </select>
-      </div>
 
-      <div style={{ display: 'flex', alignItems: 'center', gap: 6, paddingLeft: 10, borderLeft: '1px solid var(--border)' }}>
-        <button
-          onClick={onTogglePause}
-          title={isPaused ? 'Resume' : 'Pause'}
-          style={{ background: 'transparent', border: 'none', cursor: 'pointer', color: isPaused ? 'var(--orange)' : 'var(--text3)', padding: 2, display: 'flex', alignItems: 'center' }}
-        >
-          {isPaused ? <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor"><path d="M8 5v14l11-7z" /></svg> : <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor"><path d="M6 19h4V5H6v14zm8-14v14h4V5h-4z" /></svg>}
-        </button>
-        <button
-          onClick={onClear}
-          title="Clear Plot"
-          style={{ background: 'transparent', border: 'none', cursor: 'pointer', color: 'var(--red)', padding: 2, display: 'flex', alignItems: 'center' }}
-        >
-          <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round"><polyline points="3 6 5 6 21 6" /><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" /></svg>
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const AddChannelPanel = ({ boardOptions, boardLabels, boardKinds, boardColors, selectedPins, setSelectedPins, onClose }) => {
-  const boards = (boardOptions || []).filter(id => id !== 'all');
-  const [selectedBoardId, setSelectedBoardId] = React.useState(boards[0] || null);
-
-  const tabsRef = React.useRef(null);
-  const [canScroll, setCanScroll] = React.useState({ left: false, right: false });
-
-  const checkScroll = React.useCallback(() => {
-    if (!tabsRef.current) return;
-    const { scrollLeft, scrollWidth, clientWidth } = tabsRef.current;
-    setCanScroll({
-      left: scrollLeft > 2,
-      right: scrollLeft + clientWidth < scrollWidth - 2
-    });
-  }, []);
-
-  const scrollTabs = (direction) => {
-    if (!tabsRef.current) return;
-    const amount = 120;
-    tabsRef.current.scrollBy({
-      left: direction === 'left' ? -amount : amount,
-      behavior: 'smooth'
-    });
-  };
-
-  const panelRef = React.useRef(null);
-
-  React.useEffect(() => {
-    const handleClickOutside = (e) => {
-      if (panelRef.current && !panelRef.current.contains(e.target)) {
-        onClose();
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside, true);
-    return () => document.removeEventListener('mousedown', handleClickOutside, true);
-  }, [onClose]);
-
-  React.useEffect(() => {
-    checkScroll();
-    window.addEventListener('resize', checkScroll);
-    return () => window.removeEventListener('resize', checkScroll);
-  }, [checkScroll, boards.length]);
-
-  const UNO_BASE_PINS = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', 'A0', 'A1', 'A2', 'A3', 'A4', 'A5'];
-  const PICO_BASE_PINS = Array.from({ length: 29 }, (_, idx) => `GP${idx}`);
-
-  const activeBoardId = selectedBoardId || boards[0];
-  const activeKind = boardKinds[activeBoardId] || 'arduino_uno';
-  const activePins = activeKind === 'rp2040' ? PICO_BASE_PINS : UNO_BASE_PINS;
-  const activeBoardColor = boardColors[activeBoardId] || 'var(--accent)';
-
-  return (
-    <div ref={panelRef} style={{
-      position: 'absolute', top: 42, left: 12, right: 12,
-      maxHeight: 'calc(100% - 60px)',
-      background: 'var(--bg1)',
-      border: '1px solid var(--border)',
-      borderRadius: 12,
-      zIndex: 150,
-      display: 'flex',
-      flexDirection: 'column',
-      overflow: 'hidden',
-      animation: 'slideInPlotter 0.2s cubic-bezier(0,0,0.2,1)',
-      backdropFilter: 'blur(20px)',
-      boxShadow: '0 20px 50px rgba(0,0,0,0.8), 0 0 0 1px rgba(255,255,255,0.05)'
-    }}>
-      <div style={{ padding: '12px 14px', borderBottom: '1px solid var(--border)', display: 'flex', justifyContent: 'space-between', alignItems: 'center', background: 'var(--bg3)' }}>
-
-        <div style={{ display: 'flex', alignItems: 'center', gap: 8 }}>
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round" style={{ color: 'var(--accent)' }}><path d="M12 20v-6M6 20V10M18 20V4" /></svg>
-          <span style={{ fontSize: 11, fontWeight: 800, color: 'var(--text1)', textTransform: 'uppercase', letterSpacing: 0.5 }}>Add Channel</span>
-        </div>
-        <button onClick={onClose} style={{ background: 'transparent', border: 'none', color: 'var(--text3)', cursor: 'pointer', fontSize: 16 }} className="hover:text-[var(--red)] transition-colors">✕</button>
-      </div>
-
-      {/* Board Selection Bar */}
-      <div style={{
-        display: 'flex',
-        alignItems: 'center',
-        background: 'var(--bg2)',
-        borderBottom: '1px solid var(--border)',
-        height: 38,
-        position: 'relative',
-        padding: '0 4px'
-      }}>
-        {canScroll.left && (
-          <div style={{ position: 'absolute', left: 0, top: 0, bottom: 0, width: 24, zIndex: 2, background: 'linear-gradient(90deg, var(--bg2) 40%, transparent)', display: 'flex', alignItems: 'center', justifyContent: 'flex-start' }}>
-            <button onClick={() => scrollTabs('left')} style={{ background: 'none', border: 'none', padding: 0, color: 'var(--text3)', cursor: 'pointer', display: 'flex', alignItems: 'center' }}>
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><path d="m15 18-6-6 6-6" /></svg>
-            </button>
-          </div>
-        )}
-
-        <div
-          ref={tabsRef}
-          onScroll={checkScroll}
-          className="panel-scroll"
-          style={{
-            display: 'flex',
-            flex: 1,
-            overflowX: 'auto',
-            overflowY: 'hidden',
-            scrollbarWidth: 'none',
-            msOverflowStyle: 'none',
-            padding: '4px 0'
-          }}
-        >
-          {boards.map(id => (
-            <button
-              key={id}
-              onClick={() => setSelectedBoardId(id)}
-              style={{
-                flexShrink: 0,
-                padding: '4px 12px',
-                margin: '0 2px',
-                fontSize: 11,
-                fontWeight: 700,
-                color: activeBoardId === id ? 'var(--accent)' : 'var(--text3)',
-                background: activeBoardId === id ? 'rgba(0,255,255,0.08)' : 'transparent',
-                border: 'none',
-                borderRadius: 6,
-                cursor: 'pointer',
-                display: 'flex',
-                alignItems: 'center',
-                gap: 6,
-                whiteSpace: 'nowrap',
-                transition: 'all 0.2s'
-              }}
-            >
-              <span style={{
-                width: 6,
-                height: 6,
-                borderRadius: '50%',
-                background: boardColors[id] || 'var(--text4)',
-                boxShadow: activeBoardId === id ? `0 0 6px ${boardColors[id] || 'var(--text4)'}` : 'none'
-              }} />
-              {boardLabels[id] || id}
-            </button>
-          ))}
-        </div>
-
-        {canScroll.right && (
-          <div style={{ position: 'absolute', right: 0, top: 0, bottom: 0, width: 24, zIndex: 2, background: 'linear-gradient(-90deg, var(--bg2) 40%, transparent)', display: 'flex', alignItems: 'center', justifyContent: 'flex-end' }}>
-            <button onClick={() => scrollTabs('right')} style={{ background: 'none', border: 'none', padding: 0, color: 'var(--text3)', cursor: 'pointer', display: 'flex', alignItems: 'center' }}>
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><path d="m9 18 6-6-6-6" /></svg>
-            </button>
-          </div>
-        )}
-      </div>
-
-      <div style={{ flex: 1, overflowY: 'auto', padding: 14 }} className="panel-scroll">
-        {!activeBoardId ? (
-          <div style={{ padding: '20px', textAlign: 'center', color: 'var(--text3)', fontSize: 12 }}>
-            No boards detected in circuit.
-          </div>
-        ) : (
-          <div>
-            <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(42px, 1fr))', gap: 6 }}>
-              {activePins.map(pin => {
-                const pinIdx = selectedPins.findIndex(p => p.boardId === activeBoardId && p.pinId === pin);
-                const isSelected = pinIdx >= 0;
-
-                const PLOTTER_COLORS = ['#e74c3c', '#3498db', '#2ecc71', '#f1c40f', '#9b59b6', '#e67e22', '#1abc9c'];
-                const selectedColor = isSelected ? PLOTTER_COLORS[pinIdx % PLOTTER_COLORS.length] : 'var(--accent)';
-
-                return (
-                  <button
-                    key={pin}
-                    onClick={() => {
-                      setSelectedPins(prev => {
-                        if (isSelected) return prev.filter(p => p.boardId !== activeBoardId || p.pinId !== pin);
-                        if (prev.length >= 12) return prev; // Limit channels
-                        return [...prev, { boardId: activeBoardId, pinId: pin }];
-                      });
-                    }}
-                    style={{
-                      padding: '8px 2px',
-                      fontSize: 10,
-                      fontWeight: 700,
-                      background: isSelected ? `${selectedColor}25` : 'rgba(255,255,255,0.02)',
-                      color: isSelected ? selectedColor : 'var(--text3)',
-                      border: `1px solid ${isSelected ? selectedColor : 'var(--border)'}`,
-                      borderRadius: 6,
-                      cursor: 'pointer',
-                      transition: 'all 0.15s cubic-bezier(0.4, 0, 0.2, 1)',
-                      boxShadow: isSelected ? `inset 0 0 10px ${selectedColor}15` : 'none'
-                    }}
-                    className={isSelected ? '' : 'hover:border-[var(--text4)] hover:bg-white/5'}
-                  >{pin}</button>
-                );
-              })}
-            </div>
-          </div>
-        )}
-      </div>
-
-      <style>{`
-        @keyframes slideInPlotter { from { opacity: 0; transform: translateY(-8px) scale(0.98); } to { opacity: 1; transform: translateY(0) scale(1); } }
-      `}</style>
-    </div>
-  );
-};
 
 export const RightPanel = React.memo(RightPanelBase);
 
-// ── Serial UI Sub-Components ────────────────────────────────────────────────
-
-const SerialTabBar = ({
-  activeBoard,
-  otherActiveBoard,
-  setBoard,
-  isPaused,
-  onTogglePause,
-  autoscroll,
-  onToggleAutoscroll,
-  onClear,
-  onToggleSplit,
-  isSplit,
-  boardOptions,
-  boardColors,
-  boardLabels,
-  boardKinds
-}) => {
-  const boards = (boardOptions || []).filter(id => id !== 'all');
-  const tabsRef = React.useRef(null);
-  const [canScroll, setCanScroll] = React.useState({ left: false, right: false });
-  const [isHovered, setIsHovered] = React.useState(false);
-
-  const checkScroll = React.useCallback(() => {
-    if (!tabsRef.current) return;
-    const { scrollLeft, scrollWidth, clientWidth } = tabsRef.current;
-    setCanScroll({
-      left: scrollLeft > 2,
-      right: scrollLeft + clientWidth < scrollWidth - 2
-    });
-  }, []);
-
-  const scrollTabs = (direction) => {
-    if (!tabsRef.current) return;
-    const amount = 150;
-    tabsRef.current.scrollBy({
-      left: direction === 'left' ? -amount : amount,
-      behavior: 'smooth'
-    });
-  };
-
-  React.useEffect(() => {
-    checkScroll();
-    window.addEventListener('resize', checkScroll);
-    return () => window.removeEventListener('resize', checkScroll);
-  }, [checkScroll, boards.length]);
-
-  return (
-    <div
-      style={{
-        display: 'flex',
-        alignItems: 'center',
-        gap: 2,
-        padding: '4px 8px 0',
-        background: 'var(--bg2)',
-        borderBottom: '1px solid var(--border)',
-        height: 36,
-        flexShrink: 0,
-        position: 'relative'
-      }}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-    >
-      <div style={{ position: 'relative', flex: 1, display: 'flex', alignItems: 'center', overflow: 'hidden' }}>
-        {/* Left Indicator */}
-        {isHovered && canScroll.left && (
-          <button
-            onClick={() => scrollTabs('left')}
-            style={{
-              position: 'absolute', left: 0, top: 0, bottom: 0, width: 28,
-              background: 'linear-gradient(to right, var(--bg2) 60%, transparent)',
-              display: 'flex', alignItems: 'center', paddingLeft: 4, zIndex: 10,
-              border: 'none', cursor: 'pointer', color: 'var(--accent)',
-              animation: 'fadeIn 0.2s'
-            }}
-            className="hover:scale-110 transition-transform"
-          >
-            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><polyline points="15 18 9 12 15 6" /></svg>
-          </button>
-        )}
-
-        <div
-          ref={tabsRef}
-          onScroll={checkScroll}
-          style={{ display: 'flex', gap: 2, flex: 1, overflowX: 'auto', scrollbarWidth: 'none' }}
-          className="hide-scrollbar"
-        >
-          {boards.map(id => {
-            const isActive = activeBoard === id;
-            const isDisabled = otherActiveBoard === id;
-            const boardColor = boardColors[id] || '#64748b';
-            const kind = boardKinds?.[id] || 'arduino_uno';
-
-            return (
-              <button
-                key={id}
-                onClick={() => !isDisabled && setBoard(id)}
-                disabled={isDisabled}
-                style={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  gap: 6,
-                  padding: '6px 12px',
-                  fontSize: 11,
-                  fontWeight: isActive ? 700 : 500,
-                  color: isActive ? 'var(--accent)' : isDisabled ? 'var(--text4)' : 'var(--text2)',
-                  background: isActive ? 'rgba(0,255,255,0.08)' : 'transparent',
-                  border: 'none',
-                  borderBottom: isActive ? '2px solid var(--accent)' : '2px solid transparent',
-                  cursor: isDisabled ? 'not-allowed' : 'pointer',
-                  whiteSpace: 'nowrap',
-                  transition: 'all 0.2s',
-                  opacity: isDisabled ? 0.4 : 1,
-                  fontFamily: 'JetBrains Mono, monospace'
-                }}
-              >
-                <span style={{
-                  width: 7,
-                  height: 7,
-                  borderRadius: kind === 'rp2040' ? '1px' : '50%',
-                  background: boardColor,
-                  boxShadow: isActive ? `0 0 6px ${boardColor}` : 'none'
-                }} />
-                {boardLabels?.[id] || id}
-              </button>
-            );
-          })}
-        </div>
-
-        {/* Right Indicator */}
-        {isHovered && canScroll.right && (
-          <button
-            onClick={() => scrollTabs('right')}
-            style={{
-              position: 'absolute', right: 0, top: 0, bottom: 0, width: 28,
-              background: 'linear-gradient(to left, var(--bg2) 60%, transparent)',
-              display: 'flex', alignItems: 'center', justifyContent: 'flex-end', paddingRight: 4, zIndex: 10,
-              border: 'none', cursor: 'pointer', color: 'var(--accent)',
-              animation: 'fadeIn 0.2s'
-            }}
-            className="hover:scale-110 transition-transform"
-          >
-            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><polyline points="9 18 15 12 9 6" /></svg>
-          </button>
-        )}
-      </div>
-
-      <style>{`
-        @keyframes fadeIn { from { opacity: 0; } to { opacity: 1; } }
-      `}</style>
-
-      <div style={{ display: 'flex', alignItems: 'center', gap: 8, paddingLeft: 8, borderLeft: '1px solid var(--border)', marginLeft: 2 }}>
-        <button
-          onClick={() => onToggleAutoscroll(!autoscroll)}
-          style={{
-            background: autoscroll ? 'rgba(0, 255, 255, 0.08)' : 'transparent',
-            border: `1px solid ${autoscroll ? 'var(--accent)' : 'var(--border)'}`,
-            color: autoscroll ? 'var(--accent)' : 'var(--text3)',
-            padding: '3px 8px',
-            borderRadius: 4,
-            cursor: 'pointer',
-            display: 'flex',
-            alignItems: 'center',
-            gap: 5,
-            transition: 'all 0.2s',
-            userSelect: 'none',
-            marginRight: 4
-          }}
-          className="hover:border-[var(--accent)]"
-        >
-          <span style={{ fontSize: 9, fontWeight: 700, textTransform: 'uppercase', letterSpacing: 0.5 }}>Autoscroll</span>
-        </button>
-
-        <button
-          onClick={onTogglePause}
-          title={isPaused ? 'Resume' : 'Pause'}
-          style={{
-            background: 'transparent', border: 'none', cursor: 'pointer', padding: 2, borderRadius: 4,
-            color: isPaused ? 'var(--orange)' : 'var(--text3)',
-            display: 'flex', alignItems: 'center', justifyContent: 'center'
-          }}
-          className="hover:bg-white/5"
-        >
-          {isPaused ? (
-            <svg width="18" height="18" viewBox="0 0 24 24" fill="currentColor"><path d="M8 5v14l11-7z" /></svg>
-          ) : (
-            <svg width="18" height="18" viewBox="0 0 24 24" fill="currentColor"><path d="M6 19h4V5H6v14zm8-14v14h4V5h-4z" /></svg>
-          )}
-        </button>
-
-        <button
-          onClick={onClear}
-          title="Clear Output"
-          style={{
-            background: 'transparent', border: 'none', cursor: 'pointer', padding: 2, borderRadius: 4,
-            color: 'var(--red)',
-            display: 'flex', alignItems: 'center', justifyContent: 'center'
-          }}
-          className="hover:bg-[rgba(255,68,68,0.1)]"
-        >
-          <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <polyline points="3 6 5 6 21 6" /><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" /><line x1="10" y1="11" x2="10" y2="17" /><line x1="14" y1="11" x2="14" y2="17" />
-          </svg>
-        </button>
-
-        <button
-          onClick={onToggleSplit}
-          title={isSplit ? 'Single View' : 'Split View'}
-          style={{
-            background: isSplit ? 'rgba(0,255,255,0.1)' : 'transparent',
-            border: `1px solid ${isSplit ? 'var(--accent)' : 'transparent'}`,
-            cursor: 'pointer', padding: 2, borderRadius: 4,
-            color: isSplit ? 'var(--accent)' : 'var(--text3)',
-            display: 'flex', alignItems: 'center', justifyContent: 'center'
-          }}
-          className="hover:bg-white/5"
-        >
-          <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
-            <line x1="3" y1="12" x2="21" y2="12" />
-          </svg>
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const SerialOutputPane = ({ boardId, history, outputRef, isPaused, boardColors, isRunning }) => {
-  const filtered = boardId === 'all' ? history : history.filter(e => e.boardId === boardId);
-
-  return (
-    <div ref={outputRef} className="flex-1 overflow-y-auto py-1.5 flex flex-col panel-scroll" style={{ background: 'var(--bg)' }}>
-      {filtered.length === 0 ? (
-        <div style={{ color: 'var(--text3)', fontSize: 12, padding: '40px 0', textAlign: 'center', display: 'flex', flexDirection: 'column', gap: 12, alignItems: 'center' }}>
-          <span style={{ fontSize: 24, opacity: 0.5 }}>📡</span>
-          {isRunning ? `Waiting for serial output from ${boardId}...` : 'Start simulation to see serial output.'}
-        </div>
-      ) : (
-        filtered.map((entry, i) => {
-          const badgeColor = entry.dir === 'rx' ? '#2ecc71' : entry.dir === 'tx' ? '#3498db' : '#888';
-          const badgeBg = entry.dir === 'rx' ? 'rgba(46,204,113,0.12)' : entry.dir === 'tx' ? 'rgba(52,152,219,0.12)' : 'rgba(128,128,128,0.12)';
-          const boardColor = boardColors[entry.boardId] || '#64748b';
-          return (
-            <div key={i} className="flex items-start gap-2 px-3 py-0.5 text-[11px] font-mono border-b border-[var(--border)] hover:bg-white/[0.02] transition-colors">
-              <span className="text-[var(--text3)] text-[10px] min-w-[84px] shrink-0 pt-[1px]" style={{ display: 'inline-flex', alignItems: 'center', gap: 5 }}>
-                <span style={{ width: 6, height: 6, borderRadius: '50%', background: boardColor }} />
-                {entry.ts || ''}
-              </span>
-              <span className="inline-block text-[8px] font-bold rounded-[3px] px-1 py-[0px] shrink-0 mt-[2px] leading-tight" style={{ color: badgeColor, background: badgeBg, border: `1px solid ${badgeColor}40` }}>
-                {entry.dir?.toUpperCase() || 'RX'}
-              </span>
-              <span style={{ flex: 1, color: entry.dir === 'tx' ? '#3498db' : entry.dir === 'sys' ? 'var(--text3)' : 'var(--green)', whiteSpace: 'pre-wrap', wordBreak: 'break-all' }}>
-                {entry.text}
-              </span>
-              {boardId === 'all' && (
-                <span style={{ color: 'var(--text3)', fontSize: 9, minWidth: 60, textAlign: 'right', opacity: 0.7 }}>
-                  {entry.boardId || '-'}
-                </span>
-              )}
-            </div>
-          );
-        })
-      )}
-    </div>
-  );
-};
-
-const BaudRateSelector = ({ value, onChange, theme }) => {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const menuRef = React.useRef(null);
-  const options = [300, 1200, 2400, 4800, 9600, 19200, 38400, 57600, 74880, 115200, 230400, 250000, 500000, 1000000, 2000000];
-
-  React.useEffect(() => {
-    const handleClick = (e) => { if (menuRef.current && !menuRef.current.contains(e.target)) setIsOpen(false); };
-    document.addEventListener('mousedown', handleClick);
-    return () => document.removeEventListener('mousedown', handleClick);
-  }, []);
-
-  return (
-    <div style={{ position: 'relative' }} ref={menuRef}>
-      <button
-        onClick={() => setIsOpen(!isOpen)}
-        style={{
-          background: 'var(--card)', border: '1px solid var(--border)', borderRadius: 6,
-          padding: '6px 10px', fontSize: 11, color: 'var(--text2)', cursor: 'pointer',
-          display: 'flex', alignItems: 'center', gap: 6, transition: 'all 0.2s', minWidth: 80,
-          fontFamily: 'JetBrains Mono, monospace'
-        }}
-        className="hover:border-[var(--accent)]"
-      >
-        <span style={{ flex: 1, textAlign: 'left' }}>{value}</span>
-        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round" style={{ opacity: 0.5 }}>
-          <polyline points="6 9 12 15 18 9" />
-        </svg>
-      </button>
-      {isOpen && (
-        <div style={{
-          position: 'absolute', bottom: '100%', right: 0, marginBottom: 8, width: 100,
-          background: 'rgba(20, 20, 25, 0.85)', backdropFilter: 'blur(16px)',
-          border: '1px solid rgba(255, 255, 255, 0.1)', borderRadius: 8,
-          boxShadow: '0 8px 32px rgba(0, 0, 0, 0.4)', padding: 4, zIndex: 100,
-          maxHeight: 200, overflowY: 'auto'
-        }} className="hide-scrollbar">
-          {options.map(opt => (
-            <div
-              key={opt}
-              onClick={() => { onChange(opt); setIsOpen(false); }}
-              style={{
-                padding: '6px 10px', fontSize: 11, color: opt === value ? 'var(--accent)' : 'var(--text2)',
-                cursor: 'pointer', borderRadius: 4, transition: 'all 0.15s',
-                background: opt === value ? 'rgba(0, 255, 255, 0.08)' : 'transparent',
-                fontFamily: 'JetBrains Mono, monospace'
-              }}
-              className="hover:bg-white/5"
-            >
-              {opt}
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const SerialSendRow = ({
-  boardId,
-  input,
-  setInput,
-  onSend,
-  isRunning,
-  hardwareConnected,
-  serialLineEnding,
-  setSerialLineEnding,
-  serialBaudRate,
-  setSerialBaudRate,
-  boardLabels,
-  theme
-}) => {
-  return (
-    <div style={{ display: 'flex', gap: 6, padding: '8px 10px', borderTop: '1px solid var(--border)', flexShrink: 0, background: 'var(--bg2)', alignItems: 'center' }}>
-      <input
-        className="bg-[var(--card)] border border-[var(--border)] text-[var(--text)] outline-none font-inherit"
-        style={{
-          flex: 1,
-          fontFamily: 'JetBrains Mono, monospace',
-          fontSize: 11,
-          transition: 'border-color 0.2s',
-          borderRadius: 6,
-          padding: '6px 14px'
-        }}
-        placeholder={`Send to ${boardLabels?.[boardId] || boardId}...`}
-        value={input}
-        onChange={e => setInput(e.target.value)}
-        onKeyDown={e => {
-          if (e.key === 'Enter') onSend(boardId, input, serialLineEnding, serialBaudRate);
-        }}
-        disabled={!isRunning && !hardwareConnected}
-      />
-      <LineEndingSelector value={serialLineEnding} onChange={setSerialLineEnding} theme={theme} />
-      <button
-        onClick={() => onSend(boardId, input, serialLineEnding, serialBaudRate)}
-        disabled={!isRunning && !hardwareConnected}
-        style={{
-          background: (isRunning || hardwareConnected) ? 'var(--accent)' : 'var(--bg3)',
-          color: (isRunning || hardwareConnected) ? '#000' : 'var(--text4)',
-          border: 'none',
-          cursor: (isRunning || hardwareConnected) ? 'pointer' : 'not-allowed',
-          padding: '6px 20px',
-          borderRadius: 10,
-          fontSize: 11,
-          fontWeight: 800,
-          transition: 'all 0.2s',
-          fontFamily: "'Space Grotesk', sans-serif",
-          textTransform: 'uppercase',
-          letterSpacing: '0.5px'
-        }}
-        className="hover:opacity-90 active:scale-95"
-      >
-        Send
-      </button>
-    </div>
-  );
-};
-
-const LineEndingSelector = ({ value, onChange, theme }) => {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const [hoveredIdx, setHoveredIdx] = React.useState(null);
-  const menuRef = React.useRef(null);
-
-  const options = [
-    { label: 'No line ending', value: 'none' },
-    { label: 'Newline', value: 'nl' },
-    { label: 'Carriage return', value: 'cr' },
-    { label: 'Both NL & CR', value: 'crlf' }
-  ];
-
-  const currentOption = options.find(o => o.value === value) || options[1];
-
-  React.useEffect(() => {
-    const handleDown = (e) => {
-      if (menuRef.current && !menuRef.current.contains(e.target)) setIsOpen(false);
-    };
-    if (isOpen) document.addEventListener('mousedown', handleDown);
-    return () => document.removeEventListener('mousedown', handleDown);
-  }, [isOpen]);
-
-  return (
-    <div style={{ position: 'relative' }} ref={menuRef}>
-      <button
-        onClick={() => setIsOpen(!isOpen)}
-        style={{
-          background: 'var(--card)',
-          border: '1px solid var(--border)',
-          color: 'var(--text2)',
-          padding: '6px 14px',
-          fontSize: 10,
-          fontWeight: 600,
-          display: 'flex',
-          alignItems: 'center',
-          gap: 6,
-          borderRadius: 20,
-          cursor: 'pointer',
-          whiteSpace: 'nowrap',
-          transition: 'all 0.2s',
-          fontFamily: "'Space Grotesk', sans-serif"
-        }}
-        className="hover:text-[var(--accent)] hover:border-[var(--accent)]"
-      >
-        {currentOption.label}
-        <svg
-          width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3"
-          style={{ transform: isOpen ? 'rotate(180deg)' : 'none', transition: 'transform 0.2s' }}
-        >
-          <path d="M6 9l6 6 6-6" />
-        </svg>
-      </button>
-
-      {isOpen && (
-        <div style={{
-          position: 'absolute',
-          bottom: 'calc(100% + 8px)',
-          right: 0,
-          background: theme === 'light' ? 'rgba(248, 250, 252, 0.95)' : 'rgba(13, 21, 37, 0.94)',
-          backdropFilter: 'blur(16px) saturate(1.4)',
-          WebkitBackdropFilter: 'blur(16px) saturate(1.4)',
-          border: theme === 'light' ? '1px solid rgba(203, 213, 225, 0.8)' : '1px solid rgba(30, 45, 71, 0.8)',
-          borderRadius: 12,
-          boxShadow: theme === 'light' ? '0 8px 32px rgba(0, 0, 0, 0.08)' : '0 10px 40px rgba(0,0,0,0.5)',
-          zIndex: 1000,
-          minWidth: 160,
-          overflow: 'hidden',
-          padding: '5px',
-          fontFamily: "'Space Grotesk', sans-serif",
-          animation: 'serialMenuIn 0.18s cubic-bezier(0.34, 1.56, 0.64, 1)',
-          transformOrigin: 'bottom right'
-        }}>
-          {options.map((opt, idx) => (
-            <div
-              key={opt.value}
-              onClick={() => { onChange(opt.value); setIsOpen(false); }}
-              onMouseEnter={() => setHoveredIdx(idx)}
-              onMouseLeave={() => setHoveredIdx(null)}
-              style={{
-                padding: '8px 14px',
-                fontSize: 12,
-                fontWeight: value === opt.value ? 600 : 500,
-                color: value === opt.value ? 'var(--text)' : 'var(--text2)',
-                background: hoveredIdx === idx ? 'var(--bg3)' : (value === opt.value ? 'rgba(52, 152, 219, 0.1)' : 'transparent'),
-                cursor: 'pointer',
-                transition: 'all 0.15s',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'space-between',
-                borderRadius: 8,
-                margin: '1px 0'
-              }}
-            >
-              {opt.label}
-              {value === opt.value && (
-                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="var(--accent)" strokeWidth="3"><polyline points="20 6 9 17 4 12" /></svg>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-      <style>{`
-        @keyframes serialMenuIn {
-          from { opacity: 0; transform: translateY(10px) scale(0.95); }
-          to { opacity: 1; transform: translateY(0) scale(1); }
-        }
-      `}</style>
-    </div>
-  );
-};
